Rearrange routes
